command-queue: move CommandHandlerRegistryImpl to base package
