+++ core/contract/src/main/java/org/eclipse/dataspaceconnector/contract/ContractNegotiationCommandExtension.java
@@ -14,7 +14,7 @@
 package org.eclipse.dataspaceconnector.contract;
 
 import org.eclipse.dataspaceconnector.contract.negotiation.command.handlers.CancelNegotiationCommandHandler;
-import org.eclipse.dataspaceconnector.core.CommandHandlerRegistryImpl;
+import org.eclipse.dataspaceconnector.core.base.CommandHandlerRegistryImpl;
 import org.eclipse.dataspaceconnector.core.CoreExtension;
 import org.eclipse.dataspaceconnector.spi.command.CommandHandlerRegistry;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
